pyproject.toml
[tool.poetry]
name = "hydra-spl"
version = "0.1.0"
description = ""
authors = ["c2io <c2io@outlook.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
hydra-core = "^1.1.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "^21.12b0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true